requirements.txt
pandas
sqlalchemy
PyYAML
tabula_py
JPype1
re